+++ DataStore2/SavingMethods/init.lua
@@ -1,3 +1,4 @@
 return {
+	Standard = require(script.Standard),
 	OrderedBackups = require(script.OrderedBackups),
 }